refactor: address review on alert copy mappings and confirmation source
Moves the Blockaid reason-to-string maps out of the alerts utils module into
the sibling constants module, so utils holds only behavior. Replaces the
deprecated currentConfirmation read in useSendingAssetsFiatTotal with
useTransactionMetadataRequestOptional, the replacement named in the
ConfirmContextType deprecation notice.
